ADDED:[Mehul] store data in database

+++ src/main/java/com/bridgelabz/addressbook/service/IAddressBookService.java
@@ -9,11 +9,11 @@ import java.util.List;
 public interface IAddressBookService {
     ResponseEntity<List<AddressBook>> getAddressBookData();
 
-    ResponseEntity<AddressBook> getAddressBookDataById(int personId);
+    ResponseEntity<AddressBook> getAddressBookDataById(long personId);
 
     ResponseEntity<AddressBook> createAddressBookData(AddressBookDTO addressBookDTO);
 
-    ResponseEntity<AddressBook> updateAddressBookData(int personId, AddressBookDTO addressBookDTO);
+    ResponseEntity<AddressBook> updateAddressBookData(long personId, AddressBookDTO addressBookDTO);
 
-    ResponseEntity<String> deleteAddressBookData(int personId);
+    ResponseEntity<String> deleteAddressBookData(long personId);
 }
